src/main/java/ad/dummies/p01basics/c01algorithms/E04Collatz.java
package ad.dummies.p01basics.c01algorithms;

/**
 * <p>Example from the german book "Algorithms and data structures for
 * dummies":</p>
 *
 * <p>A. Gogol-Döring and T. Letschert, <i>Algorithmen und Datenstrukturen für
 * Dummies</i>. Weinheim, Germany: Wiley-VCH, 2019.</p>
 *
 * <p>The current version of these examples with unit tests and benchmarks can
 * be found <a href="https://github.com/CSchoel/ad-dummies-java">on GitHub</a>.
 * </p>
 *
 * @author Christopher Schölzel
 */
public class E04Collatz {
    public static void collatz(int n) {
        while (n > 1) {
            System.out.println(n);
            if (n % 2 == 0) {
                n /= 2;
            } else {
                n = 3 * n + 1;
            }
        }
        System.out.println(n); // always 1
    }

    public static void main(String[] args) {
        collatz(25);
    }
}
